transfer js from home.vue to home.js and put the api key and secret key in env
