+++ .env
@@ -1,1 +1,3 @@
-VITE_URL = "http://192.168.3.23:8002"
+VITE_URL = "http://192.168.3.23:8000"
+VITE_API_KEY = "a8f68ca349940e1"
+VITE_API_SECRET = "5ac9be99799c90f"
