+++ .gitignore
@@ -7,6 +7,10 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+# Environment variables
+#.env
+#.env.*
+
 node_modules
 .DS_Store
 dist
